refactor(app): group routing into src/app/routing
Every loose file in src/app except app-info.ts was routing — the
router, route table, patterns, error boundary — while loaders/ sat
beside them despite serving only the route table. Move them all under
src/app/routing so the folder reads one-per-concept like its siblings.

Repoint the loader-scoped ESLint only-throw-error override and the
architecture doc paths that referenced src/app/loaders.
