+++ src/app/routing/app-routes.tsx
@@ -1,9 +1,9 @@
 import { AppShell } from "@app/layouts/app-shell";
-import { boardLoader } from "@app/loaders/board-loader";
-import { boardSetIndexLoader } from "@app/loaders/board-set-index-loader";
-import { rootIndexLoader } from "@app/loaders/root-index-loader";
-import { RouteErrorBoundary } from "@app/route-error-boundary";
-import { ROUTE_PATTERNS } from "@app/route-patterns";
+import { boardLoader } from "@app/routing/loaders/board-loader";
+import { boardSetIndexLoader } from "@app/routing/loaders/board-set-index-loader";
+import { rootIndexLoader } from "@app/routing/loaders/root-index-loader";
+import { RouteErrorBoundary } from "@app/routing/route-error-boundary";
+import { ROUTE_PATTERNS } from "@app/routing/route-patterns";
 import { BOARD_PATTERN, BOARD_SET_PATTERN } from "@features/board";
 import { LoadingState } from "@shared/components/loading-state";
 import { type RouteObject } from "react-router";
