+++ src/main.tsx
@@ -1,9 +1,9 @@
-import { AppRouter } from "@app/app-router";
-import { AppProviders } from "@shared/providers/app-providers";
+import { AppRouter } from "@app/routing/app-router";
 import "@fontsource/roboto/300.css";
 import "@fontsource/roboto/400.css";
 import "@fontsource/roboto/500.css";
 import "@fontsource/roboto/700.css";
+import { AppProviders } from "@shared/providers/app-providers";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
